Refactoring
 * Moved GUI code from swingworker to worker creator code
